remove nearly all remaining deprecations (JuliaLang/julia#28486)
* Fix at-inferred with a single dot-broadcast op

* Mark .op type-inference-based completion of method signatures as broken

REPLCompletions allows tab-completion to show the possible signatures of a given function when in the midst of writing its arguments. It uses inference to determine the possible types of the already-written arguments, subsetting the possible signatures. With the removal of `.op` as a function, it is no longer able to infer these types and thus no longer subsets the signatures as previously expected. The failure mode is very graceful here -- it just assumes `Any` and shows more signatures. This can be re-implemented later