ml-matches: update and improve ambiguity computation (JuliaLang/julia#36962)
With the improved internal implementation of ambiguity testing, it's useful now to lean on that existing support more heavily and compute a lighter version here, fixing some issues with the previous one. Most of these new failures are likely to be type-intersection issues (JuliaLang/julia#36951), which previously were excluded from analysis entirely.